ProdutoService criado

+++ src/main/java/br/edu/unicesumar/trabalho/repository/ProdutoRepository.java
@@ -2,6 +2,7 @@ package br.edu.unicesumar.trabalho.repository;
 
 import br.edu.unicesumar.trabalho.entity.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 }
